+++ animatch-backend/src/models/userReactionModel.js
@@ -1,5 +1,14 @@
 import supabase from "../config/databaseConfig.js";
 
+/* This model handles user reactions to anime. Whether they like, dislike, not interested, or no interaction yet
+    A user is able to create, update, fetch, and delete their reactions to anime.
+     */
+
+/*
+    Get User is in each function to ensure the user is valid and authenticated before performing any operations.
+*/
+
+//This object helps map the types of reactions a user can have to anime to the ENUM stored in the database.
 const UserReaction_reaction = Object.freeze({
     LIKED: 'liked',
     DISLIKED: 'disliked',
@@ -9,10 +18,10 @@ const UserReaction_reaction = Object.freeze({
 
 export const fetchUserReactions = async () => {
     const {data: userData, error: userError} = await supabase.auth.getUser();
-
     if (!userData) throw new Error('User not authenticated');
     if (userError) throw userError;
 
+    // Fetch user reactions from the user_reactions table
     const {data: reactions, error: fetchError} = await supabase.from('user_reactions').select().eq('user_id', userData.user.id);
 
     if (fetchError) throw fetchError;
@@ -24,10 +33,10 @@ export const fetchUserReactions = async () => {
 
 export const fetchUserReactionsWithAnimeTitles = async () => {
     const {data: userData, error: userError} = await supabase.auth.getUser();
-
     if (!userData) throw new Error('User not authenticated');
     if (userError) throw userError;
 
+    // Fetch user reactions along with the assoaciated anime titles from the kitsu_anime_data table
     const {data: reactions, error: fetchError} = await supabase
     .from('user_reactions')
     .select(`anime_id, reaction, kitsu_anime_data (id, title)`)
@@ -35,7 +44,6 @@ export const fetchUserReactionsWithAnimeTitles = async () => {
 
 
     if (fetchError) throw fetchError;
-
     if (!reactions) throw new Error('No reactions found for user');
 
     return reactions;
@@ -43,18 +51,17 @@ export const fetchUserReactionsWithAnimeTitles = async () => {
 
 export const createUserReaction = async ({anime_id, reaction}) => {
     const {data: userData, error: userError} = await supabase.auth.getUser();
-
     if (!userData) throw new Error('User not authenticated');
     if (userError) throw userError;
 
 
+    // Validate the reaction type
     if (!Object.values(UserReaction_reaction).includes(reaction)) throw new Error('Invalid reaction type');
     else reaction = UserReaction_reaction[reaction.toUpperCase()];
 
     const response = await supabase.from('user_reactions').insert({anime_id: anime_id, reaction, user_id: userData.user.id}).select();
 
     if(!response || response.length == 0) throw new Error('Failed to create user reaction');
-
     if (response.error) throw response.error;
 
 
@@ -62,30 +69,37 @@ export const createUserReaction = async ({anime_id, reaction}) => {
 }
 
 export const updateUserReaction = async (reactionId, updatedData) => {
-    const {data: existingReaction, error: fetchError} = await supabase.from('user_reactions').select().eq('id', reactionId);
+    const {data: userData, error: userError} = await supabase.auth.getUser();
+    if (!userData) throw new Error('User not authenticated');
+    if (userError) throw userError;
 
+    //check if the specific user reaction exists before trying to update it
+    const {data: existingReaction, error: fetchError} = await supabase.from('user_reactions').select().eq('id', reactionId);
     if (!existingReaction) throw new Error('User reaction does not exist');
     if (fetchError) throw fetchError;
 
     const response = await supabase.from('user_reactions').update(updatedData).eq('id', reactionId).select();
 
     if (!response || response.length === 0) throw new Error('Failed to update user reaction');
-
     if (response.error) throw response.error;
 
     return response;
 }
 
 export const deleteUserReaction = async (reactionId) => {
-    const {data: existingReaction, error: fetchError} = await supabase.from('user_reactions').select().eq('id', reactionId);
+    const {data: userData, error: userError} = await supabase.auth.getUser();
+    if (!userData) throw new Error('User not authenticated');
+    if (userError) throw userError;
 
+    //check if the specific user reaction exists before trying to delete it
+    const {data: existingReaction, error: fetchError} = await supabase.from('user_reactions').select().eq('id', reactionId);
     if (!existingReaction) throw new Error('User reaction does not exist');
     if (fetchError) throw fetchError;
+
 
     const response = await supabase.from('user_reactions').delete().eq('id', reactionId);
 
     if(!response || response.length === 0) throw new Error('Failed to delete user reaction');
-
     if (response.error) throw response.error;
 
     return response;
